Refactored file architecture and namespaces - DataAccess project now has everything inside Models, Interfaces, Repositories folders. - Domain project was untouched. - The IHub folder and its contents were deleted because they were unused in the WebApi project. - The Tests project now has everything inside UnitTests, IntegrationTests folder. - The namespaces of the files match the architecture.
