+++ FakebookNotifications/FakebookNotifications.DataAccess/Models/NotificationsContext.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using FakebookNotifications.DataAccess.Models;
-using FakebookNotifications.Domain.Interfaces;
+using FakebookNotifications.DataAccess.Interfaces;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using MongoDB.Bson;
 using MongoDB.Driver;
 
-namespace FakebookNotifications.DataAccess
+namespace FakebookNotifications.DataAccess.Models
 {
     public class NotificationsContext : INotificationsContext
     {
