+++ FakebookNotifications/FakebookNotifications.DataAccess/Repositories/UserRepo.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+using FakebookNotifications.DataAccess.Interfaces;
 using FakebookNotifications.DataAccess.Models;
 using FakebookNotifications.Domain.Interfaces;
 using MongoDB.Driver;
 
-namespace FakebookNotifications.DataAccess
+namespace FakebookNotifications.DataAccess.Repositories
 {
     public class UserRepo : IUserRepo
     {
