+++ FakebookNotifications/FakebookNotifications.Testing/IntegrationTests/UserRepoTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
-using FakebookNotifications.DataAccess;
 using FakebookNotifications.DataAccess.Models;
+using FakebookNotifications.DataAccess.Repositories;
 using FakebookNotifications.Domain.Interfaces;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
@@ -8,7 +8,7 @@ using MongoDB.Driver;
 using Moq;
 using Xunit;
 
-namespace FakebookNotifications.Testing
+namespace FakebookNotifications.Testing.IntegrationTests
 {
     public class UserRepoTests
     {
